runtime/autoload/netrw.vim: Allow for empty paths
URIs such as `https://hashbang.sh` are valid, but netrw does not
detect them properly, and throws an error, as it expects a path. A
blank path can be treated as "/", so this fix checks if the path
component is missing, and if it is, appends a "/" onto the path.